Fixed BOSH polling behavior bug introduced in r11555. See: http://www.igniterealtime.org/community/thread/40903?tstart=0
Problem was that connection was not sending new empty request if there were no responses from the server, because resetResponseProcessor() will only cause a call to processResponse() if there are responses to process.  And, since the pollServer() code is now in processResponse() as a result of r11555, need to make sure to send a poll in onRequestComplete() if there are no responses.

git-svn-id: http://svn.igniterealtime.org/svn/repos/xiff/trunk@11631 b35dd754-fafc-0310-a699-88a17e54d16e